Huge metadata update.
Added lots of interfaces to the system.
The metadata system is working already right out of the box,
but there are no good implementations for tiles, fields, terrain and
terrain generation yet.
Don't mess up with the new package 'entity', since it contains
the interface 'IEntity' and the subclass 'Entity'. This can
cause ambiguous references if we don't watch out.