Huge new world update.
Generation of worlds can now rely on brushes!
The painting onto an image is still buggy and can
throw exceptions here and there, but I'm going to fix
bugs asap.
Tile system is working (it's still alpha, but hey); field system
is not implemented at all, just the basic structure.