Third huge update!
Tiles are being displayed now in isometric view. The default world size is 45x45 for the time being.
A new (temporary) screen class has been created for testing the new world API. If you want to switch back to
the old game screen, simply go to the method "initializeScreens()" in GameWindow and do the necessary changes.
A rudimentary implementation of the new entity API is provided under the new package "entity".

The map can be moved with the arrow keys; an ability to zoom is provided with the PageUp and PageDown key.
The performance of the new tile system is, especially at the drawing, at an all-time low. The point is: in class Tile
is no "updateGUI()" method yet. So the coordinates are recalculated at every drawing process.
The tile is not compatible with the "Component" superclass yet, but it will soon, as I go fix all bugs in the new tile system.