Tweaks to generation.
The performance of the drawing still sucks, but we'll
have to deal with it at a later time.
The world size has been set to 100x100 with setting the colors
of the map to green and blue (resembling land tiles and water tiles).

Sometimes it can happen that black tiles appear inside of the map.
Ignore this bug, I am going to fix it.
But the rest works perfectly fine.