PreWindowScreen (without EventHandling), Component classes and ArrowSelection(Screen) improvement
The new class PreWindowScreen is added. It will replace PreWindow completely when it's ready (as well as ArrowSelectionScreen ArrowSelection). In this untested form of PreWindowScreen (and ArrowSelectionScreen), there is no guarantee that the design will be good. PreWindowScreen does not have Event Handling right now. The audio-clip "introBackground.WAV" in resources/sfx will be the title background melodie of Pfeile (we can change it later). ComboBox got the method setSelectedIndex(int index) for setting the index after initialising a new ComboBox.