Readded the world classes.
The world classes are back up, with less code and less overhead than ever... The things that are missing right now is access to the "Component" part of the Tile and the brush classes.
But OMG THAT PERFORMANCE THOUGH! Everything runs fluid while hovering over the tiles with the mouse!

I solved the problem where to put the WorldViewport object, and the solution is called "VisualMap". Just peek into the class and you'll get it. Maybe.