Pathfinding is working.
After so long, I got pathfinding working.
The setup is as follows: the player has two movement points right now. Walking to a grass tile only costs one movement point, while walking on a sea tile
requires two movement points. This can be later adjusted of course.
Click on a tile and it will calculate the best way to the tile possible; click the "End turn" button to make him move again.
For now, the active player is just a blue box.