Updated spinner model and implemented brushes.
The spinner model has been extracted to an interface for easier use. The spinner still only supports the integer calculation model, not the generic model; I am planning to integrate the other
type as well.
PfeileContext is not up and running yet, the whole game data is still located in GameScreen (but not for long anymore).
In the mathematics section I added the vector classes as well as a preliminary implementation of a Beziér curve.
For memory management I created the VolatileResource class, which deallocates (under certain circumstances) the underlying resource automatically after a certain amount of time.