Code base change.
The PfeileContext class takes now care of the data instead of the Mechanics class. PreWindowScreen needs to be rewritten to the new class.
A new UI element has arrived, the (progress) bar.
The Life class does not contain the UI for itself in the class, but the LifeUI class does now. I am explaning the reasons later in doc comments.