Attack mechanism implemented.
The implementation may be a bit clunky, but it works for now. I am trying to make this implementation easier to use.
I tested this implementation with the arrows that are being added to the inventory of the player; result was that the arrows are being removed from the inventory and that the arrow
impacted a few turns later.

So this is done.