WanderingTraderUI (component reactes to world movements etc.)
Every wandering trader has an inner class WanderingTraderUI extends ImageComponent, which is contains the graphic interpretation of a wandering trader. WanderingTraderUI are now bound to a tile and moves, if the world is shifted, etc. 
EntityManager containes two new fields player and opponent, for an faster access and to avoid some ConcurrentModificationExceptions (it can still be thrown, if there are lot's of entities created at once and/or they search the entity list of entity manager at the same time)